Replace outdated terms (#1174)
## Changes

Replaced occurrences of "game" with "experience" and "friends" with
"connections" to match current Roblox terminology

## Checks

By submitting your pull request for review, you agree to the following:

- [ ] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [ ] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [ ] To the best of my knowledge, all proposed changes are accurate.

---------

Co-authored-by: Andrew Etter <aetter@users.noreply.github.com>