Replace “Friends” with “Connections” (#1173)
## Changes

Replaced all mentions of “Friends” with “Connections” to stay consistent
with the new naming used on the Roblox platform.

## Checks

By submitting your pull request for review, you agree to the following:

- [ ] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [ ] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [ ] To the best of my knowledge, all proposed changes are accurate.

Co-authored-by: Andrew Etter <aetter@users.noreply.github.com>