Update GuiService Documentation to Fix Mismatched Reference (#1223)
Fixes mistake where `GuiService.SelectionOrder` was referenced instead
of `GuiObject.SelectionOrder`

## Changes

`GuiService.SelectionOrder` obviously doesn't exist, i am assuming that
it was originally supposed to say `GuiObject.SelectionOrder`.
This request fixes this issue

When i came across this, i was decently confused by this at first then
after a couple of minutes i realized it was a mistake which was a little
annoying...

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

Co-authored-by: IgnisRBX <43388550+IgnisRBX@users.noreply.github.com>