Add references to FreeCam camera tilting keybinds (#1235)
## Changes

- Added respective references to the key binds for the camera tilting
feature, which can be accessed by the `Z` and `C` keys while actively
inside of freecam.

## Checks

By submitting your pull request for review, you agree to the following:

- [X] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [X] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

---------

Co-authored-by: IgnisRBX <43388550+IgnisRBX@users.noreply.github.com>