Update creator docs to reflect improved sleep system (#1242)
## Changes

Update the creator docs to reflect recent improvements to the sleep
system. This [forum
post](https://devforum.roblox.com/t/improved-sleep-system-enhancing-accuracy-consistency-and-performance/3587885)
details the improvements.

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

---------

Co-authored-by: IgnisRBX <43388550+IgnisRBX@users.noreply.github.com>