Changes to link consistencies & Inspire program status (#1247)
## Changes

- Updated legacy links from `.com/groups/` to `.com/communities/` for
consistency with current Roblox site link formats.
- Updated Inspire creator program to be marked as closed due to
applications not being open anymore.

## Checks

By submitting your pull request for review, you agree to the following:

- [X] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [X] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [X] To the best of my knowledge, all proposed changes are accurate.

---------

Co-authored-by: Andrew Etter <aetter@users.noreply.github.com>