Add clarification to buffer.create (#1237)
## Changes

Specifies the size format.

## Validation

The size format is never specified, I figured the following from my own
tests.
Negative integer: Error while creating buffer.
Positive double: The number is floored, and the extra bits are ignored.

## Checks

By submitting your pull request for review, you agree to the following:

- [X] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [X] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [X] To the best of my knowledge, all proposed changes are accurate.

---------

Co-authored-by: Andrew Etter <aetter@users.noreply.github.com>