SocialService: Fix typos in PromptLinkSharing Parameters (#1265)
## Changes

This PR fixes two typos in the `options` parameters for
PromptLinkSharing; `PreiviewDescription` and `PreviewAssettId`.

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

Co-authored-by: Andrew Etter <aetter@users.noreply.github.com>