Update matchmaking/attributes-and-signals.md (#1272)
## Changes

Issues I missed in #1271

- Fix formatting for line 229
- Instead of `error`, use `errorMessage`, as this collides with the
[error](https://create.roblox.com/docs/reference/engine/globals/LuaGlobals#error)
global (per
[pcall](https://create.roblox.com/docs/reference/engine/globals/LuaGlobals#pcall)
example)
- Fixed invalid syntax (trying to assign to variable inside if
statement)

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

---------