Fix typos in CompressorSoundEffect (#1279)
Corrected typos in the CompressorSoundEffect documentation.

## Changes

<!-- Please summarize your changes. -->
fixed typos in
\content\en-us\reference\engine\classes\CompressorSoundEffect.yaml
i am not too sure whether to PR judging off the header of the text, but
hey, its worth a shot.
<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.