Add concurrent video playback limitation
Currently, there is a warning that will be output when you try to play more than 2 videos at the same time. This is a blocker for most interesting usage of this feature (e.g. Times Square style billboards, video ads in bus stops, etc.)

Given that this feature is new-ish and might still be subject to fine-tuning, I'm not sure if it's valid to document these limits. Feel free to reject this change if these limits are still actively being assessed.