Add concurrent video playback limitation (#1285)
## Changes

Currently, there is a warning that will be output when you try to play
more than 2 videos at the same time. This is a blocker for some
interesting uses of this feature (e.g. Times Square style billboards,
video ads in bus stops, etc.) that developers should know ahead of time.

<img width="371" height="68" alt="image"
src="https://github.com/user-attachments/assets/ec08a25d-2ebb-464a-a929-b7a8e8a41d0f"
/>

Given that this feature is new-ish and might still be subject to
fine-tuning, I'm not sure if it's valid to document these limits. Feel
free to reject this change if these limits are still actively being
assessed.

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

---------

Co-authored-by: IgnisRBX <43388550+IgnisRBX@users.noreply.github.com>