Correct RunService context rest results table (#1302)
## Changes

Incorrectly handled some data when creating #1300, which I've only
noticed after it was merged, sorry!
This PR fixes those issues.

- Team Test (Server) had IsStudio as false
- Run Mode had the same data as Edit

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.