remove non-existing method related to GroupService inside Player docuemnt (#1311)
## Changes

<!-- Please summarize your changes. -->

in this change, the misleading implication that `Player:IsInGroup` is
tied to a non existing `GroupService:PromptJoinAsync` was removed.

<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

Co-authored-by: IgnisRBX <43388550+IgnisRBX@users.noreply.github.com>