Remove use of 'game.Workspace' (#1317)
## Changes

- Replaced the singular use of `game.Workspace` in
**in-experience-creation.md** with `workspace`
- Removed an info alert that at this point is useless

**Question:**
Should all use of `game:GetService("Workspace")` be removed? It is
redundant as the
[workspace](https://create.roblox.com/docs/reference/engine/globals/RobloxGlobals#workspace)
global exists. The use of the GetService method here just adds a new
top-level variable where the only difference is capitalization, while
both point to the same instance.

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.