Add Class.TextService to Class.Chat deprecation message (#1321)
## Changes

`Chat` service was split into 2 services, `TextService` and
`TextChatService`.
The current `Chat` service deprecation message only links to
`TextChatService`.

For example, the `Chat.FilterStringAsync` method is not present in
`TextChatService`, but it is in `TextService`.

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

---------

Co-authored-by: IgnisRBX <43388550+IgnisRBX@users.noreply.github.com>