Fix API references on Passes article (#1333)
## Changes
Fixed references to `PromptGamePassPurchase()` &
`PromptGamePassPurchaseFinished` on the
[Passes](https://create.roblox.com/docs/production/monetization/game-passes)
article. :^)

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.