Update "tag-table.md" to better reflect engine behavior (#1340)
## Changes

- Fixed several grammatical errors and changed some sentences throughout
descriptions and given advice to be clearer.
- Renamed "AI/navigation" and "Animation" to "AI/Navigation" and
"Animation and Humanoids" respectively
- Added new tags (Audio: Sound/stepInstances, Rendering: Shadows, UI:
UpdateUILayouts/Rebuild Z-order list, UpdateUILayouts/Layout)
- Removed the sentence on the Prepare job blocking simulation threads as
it is no longer relevant. The desktop client was the only version that
had this behavior which was recently removed.
- Removed `DeveloperTag_Sounds` and `DeveloperTag_StreamingSounds` in
the Audio section as it is only relevant in the Developer Console and
never mentioned anywhere in the MicroProfiler.

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

---------

Co-authored-by: IgnisRBX <43388550+IgnisRBX@users.noreply.github.com>
Co-authored-by: Andrew Etter <aetter@users.noreply.github.com>